+++ azure/storage/models.py
@@ -408,14 +408,14 @@ class AccessPolicy(object):
             been specified in an associated stored access policy. Azure will always 
             convert values to UTC. If a date is passed in without timezone info, it 
             is assumed to be UTC.
-        :type expiry: date or str
+        :type expiry: datetime.datetime or str
         :param start:
             The time at which the shared access signature becomes valid. If 
             omitted, start time for this call is assumed to be the time when the 
             storage service receives the request. Azure will always convert values 
             to UTC. If a date is passed in without timezone info, it is assumed to 
             be UTC.
-        :type start: date or str
+        :type start: datetime.datetime or str
         '''
         self.start = start
         self.expiry = expiry