+++ tests/settings_fake.py
@@ -18,6 +18,8 @@
 
 STORAGE_ACCOUNT_NAME = "storagename"
 STORAGE_ACCOUNT_KEY = "NzhL3hKZbJBuJ2484dPTR+xF30kYaWSSCbs2BzLgVVI1woqeST/1IgqaLm6QAOTxtGvxctSNbIR/1hW8yH+bJg=="
+BLOB_STORAGE_ACCOUNT_NAME = "blobstoragename"
+BLOB_STORAGE_ACCOUNT_KEY = "NzhL3hKZbJBuJ2484dPTR+xF30kYaWSSCbs2BzLgVVI1woqeST/1IgqaLm6QAOTxtGvxctSNbIR/1hW8yH+bJg=="
 REMOTE_STORAGE_ACCOUNT_NAME = "remotestoragename"
 REMOTE_STORAGE_ACCOUNT_KEY = "3pJwX7wjxoQEW2nKAhJZARpQWpKvPOUN9JwoV/HhlMmJlS1pORhzzHpPfQqgFcwGsriu6dwYqnugWOjGShC5VQ=="
 PREMIUM_STORAGE_ACCOUNT_NAME = "premiumstoragename"
@@ -25,6 +27,7 @@ PREMIUM_STORAGE_ACCOUNT_KEY = "NzhL3hKZbJBuJ2484dPTR+xF30kYaWSSCbs2BzLgVVI1woqeS
 
 # Use instead of STORAGE_ACCOUNT_NAME and STORAGE_ACCOUNT_KEY if custom settings are needed
 CONNECTION_STRING = ""
+BLOB_CONNECTION_STRING = ""
 PREMIUM_CONNECTION_STRING = ""
 
 # Use 'https' or 'http' protocol for sending requests, 'https' highly recommended